Added JS functionality + re-structure HTML
Added the Javascript functionality to scroll to a specific part of the page when clicked and to change the colour of the navigation bar when scrolling.

The HTML page was also restructured to include the form, links to pages and the amount raised so far.

Finally the CSS contains the physical changes made as a result of the aforementioned